+++ Common/CMakeLists.txt
@@ -2,33 +2,33 @@
 cmake_minimum_required(VERSION 3.8)
 
 set(SOURCES
-   src/BitmapImage.cpp
+   src/BitmapLoader.cpp
    src/CommandLineParser.cpp
    src/Image.cpp
    src/Pixel.cpp
    src/PrintMessageHandler.cpp
-   src/RAWImage.cpp
+   src/RAWLoader.cpp
    src/Utils.cpp
    src/Window.cpp
    src/png/Adler32.cpp
    src/png/BitStream.cpp
    src/png/HuffmanDecoder.cpp
    src/png/Inflate.cpp 
-   src/png/PngImage.cpp
+   src/png/PngLoader.cpp
    src/png/Type01Decoder.cpp
    src/png/Type10Decoder.cpp
 )
 
 set(INCLUDES
    include/ArgumentHandler.hpp
-   include/BitmapImage.hpp
+   include/BitmapLoader.hpp
    include/CommandLineParser.hpp
    include/Common.hpp
    include/Exception.hpp
    include/Image.hpp
    include/Pixel.hpp
    include/PrintMessageHandler.hpp
-   include/RAWImage.hpp
+   include/RAWLoader.hpp
    include/Utils.hpp
    include/Window.hpp
    include/png/Adler32.hpp
@@ -36,7 +36,7 @@ set(INCLUDES
    include/png/BlockDecoder.hpp
    include/png/HuffmanDecoder.hpp
    include/png/Inflate.hpp 
-   include/png/PngImage.hpp
+   include/png/PngLoader.hpp
    include/png/Type01Decoder.hpp
    include/png/Type10Decoder.hpp
 )
