+++ Common/include/BitmapLoader.hpp
