+++ Common/include/png/PngLoader.hpp
