+++ Common/src/RAWLoader.cpp
